services/data_service/requirements.txt
gunicorn==20.0.0  # Gunicorn cho môi trường UNIX (Linux)
waitress==2.1.2  # Waitress nếu bạn sử dụng trên Windows thay vì Gunicorn
marshmallow==3.14.1
python-dotenv==0.20.0
psycopg2==2.9.10
Flask==3.1.0
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
# pyspark==3.1.1
numpy==2.2.0
findspark==1.4.2
confluent-kafka==2.6.1
google-generativeai==0.8.3
langdetect==1.0.9